Removed the dynamic import system
The dynamic import is neat, but it's a pita when using code completion
to handle automatic imports. Dealing with that would make things way
more complicated, so just got rid of it and hand implemented all
elements listed on the MDN. Custom elements should be created manually
by extending Element or VoidElement.